Add business logic for login screens
Most of the screens require the user to be logged in and we don't want to replicate the logic all over.

Based on the flutter principle "everything is a widget", we are moving this business logic to LoggedScreen. This Widget gets a child widget in the constructor and returns it if the user is logged in otherwise returns the login screen. The login status is accessed via Provider<Login> which is the app root widget.

In order to create a loggedin screen it's sufficient to wrap it with a logged_screen or with a rousseau_logged_scaffold which also gives the toolbar.